+++ src/gui/vm_protobuf_message_list.rs
@@ -1,6 +1,9 @@
+use std::collections::BTreeMap;
+
 use super::vm_demo_file::tick_to_time_string;
 use super::{ Event, ViewModel, Focusable, table_constants };
-use eframe::egui::{ self, RichText, Sense, CursorIcon };
+use eframe::egui::{ self, RichText, Sense, CursorIcon, Layout };
+use eframe::emath::Align;
 use egui_extras::{ Column, TableBuilder };
 use source_demo_tool::protobuf_message::ProtobufMessageEnumTraits;
 use super::vm_protobuf_message::ProtobufMessageViewModel;
@@ -20,7 +23,11 @@ const MESSAGE_DETAIL_MIN_WIDTH: f32 = 420.0;
 
 pub struct ProtobufMessageListViewModel<MessageType: ProtobufMessageEnumTraits> {
     pub vm_protobuf_message: Option<ProtobufMessageViewModel>,
-    pub messages: Vec<MessageParseReturn<MessageType>>,
+    pub messages: BTreeMap<
+        &'static str,
+        BTreeMap<usize, MessageParseReturn<MessageType>>
+    >,
+    display_messages: Vec<(usize, MessageParseReturn<MessageType>)>,
     active_message: Option<usize>,
     name: &'static str,
     b_scroll_next: bool,
@@ -29,20 +36,85 @@ pub struct ProtobufMessageListViewModel<MessageType: ProtobufMessageEnumTraits> 
     >>,
     message_ticks: Option<Vec<i32>>,
     tick_interval: Option<f32>,
+    filterable: bool,
+    active_filter_index: usize,
+    filterable_data: Vec<(&'static str, usize)>,
+    filterable_list: Vec<String>,
 }
 
 impl<MessageType: ProtobufMessageEnumTraits + Clone + 'static> ProtobufMessageListViewModel<MessageType> {
-    pub fn new(name: &'static str, messages: Vec<MessageParseReturn<MessageType>>) -> Self {
+    pub fn new(name: &'static str, messages_vec: Vec<MessageParseReturn<MessageType>>) -> Self {
+        let mut filterable_data = BTreeMap::new();
+        let mut messages:
+            BTreeMap<
+                &'static str,
+                BTreeMap<usize, MessageParseReturn<MessageType>>
+            > = BTreeMap::new();
+
+        messages.insert("None", BTreeMap::new());
+
+        for i in 0..messages_vec.len() {
+            let msg_return = &messages_vec[i];
+            if let Some(msg) = &msg_return.message {
+                let msg_str = msg.to_str();
+                filterable_data
+                    .entry(msg_str)
+                    .and_modify(|v| {*v += 1})
+                    .or_insert(1 as usize);
+
+                messages
+                    .entry("None")
+                    .and_modify(|m| {
+                        m.insert(i, msg_return.clone());
+                    });
+
+                messages
+                    .entry(msg_str)
+                    .and_modify(|m| {
+                        m.insert(i, msg_return.clone());
+                    })
+                    .or_insert({
+                        let mut m = BTreeMap::new();
+                        m.insert(i, msg_return.clone());
+                        m
+                    });
+            }
+        }
+        let filterable_data: Vec<(&'static str, usize)> = filterable_data.into_iter().collect();
+
+        let mut filterable_list = Vec::new();
+        filterable_list.push("None".to_owned());
+        for fd in &filterable_data {
+            filterable_list.push(
+                format!(
+                    "{} ({})",
+                    fd.0, fd.1
+                )
+            );
+        }
+
+        let display_messages
+            = messages["None"].clone().into_iter().collect();
+
         Self {
             name,
             messages,
+            display_messages,
+            filterable_data,
+            filterable_list,
             vm_protobuf_message: None,
             active_message: None,
             b_scroll_next: true,
             message_header_callback: None,
             message_ticks: None,
             tick_interval: None,
+            filterable: false,
+            active_filter_index: 0,
         }
+    }
+
+    pub fn set_filterable(&mut self, filterable: bool) {
+        self.filterable = filterable;
     }
 
     pub fn set_tick_column(&mut self, ticks: Vec<i32>, tick_interval: f32) {
@@ -60,68 +132,76 @@ impl<MessageType: ProtobufMessageEnumTraits + Clone + 'static> ProtobufMessageLi
     }
 
     pub fn set_active_message(&mut self, index: usize) -> bool {
-        if index >= self.messages.len() {
-            return false
+        for msg in &self.display_messages {
+            if msg.0 == index {
+                self.b_scroll_next = true;
+                self.active_message = Some(index);
+                let active_message = &msg.1.message;
+                if let Some(msg) = active_message {
+                    self.vm_protobuf_message = Some(
+                        ProtobufMessageViewModel::new(Box::new(msg.clone()))
+                    );
+                }
+                return true
+            }
         }
-
-        self.b_scroll_next = true;
-
-        self.active_message = Some(index);
-        let active_message = &self.messages[self.active_message.unwrap()].message;
-        let active_message = active_message.as_ref().unwrap().clone();
-        self.vm_protobuf_message = Some(
-            ProtobufMessageViewModel::new(Box::new(active_message))
-        );
-        true
+        false
     }
 
     pub fn next_message(&mut self) -> bool {
         match &self.active_message {
-            Some(index) => {
-                self.set_active_message(index + 1)
-            },
-            None => {
-                if self.messages.is_empty() {
-                    false
-                } else {
-                    self.set_active_message(0);
-                    true
+            Some(_) => {
+                let mut b_capture_next = false;
+                let mut next_index = None;
+                for msg in &self.display_messages {
+                    if b_capture_next {
+                        next_index = Some(msg.0);
+                        break
+                    }
+                    if msg.0 == self.active_message.unwrap() {
+                        b_capture_next = true
+                    }
                 }
-            }
+                if let Some(i) = next_index {
+                    self.set_active_message(i)
+                } else {
+                    false
+                }
+            },
+            None => self.first_message()
         }
     }
 
     pub fn prev_message(&mut self) -> bool {
         match &self.active_message {
-            Some(index) => {
-                if *index == 0 {
-                    false
+            Some(_) => {
+                let mut prev_index = None;
+                for msg in &self.display_messages {
+                    if msg.0 == self.active_message.unwrap() {
+                        break
+                    }
+                    prev_index = Some(msg.0);
+                }
+                if let Some(i) = prev_index {
+                    self.set_active_message(i)
                 } else {
-                    self.set_active_message(index - 1);
-                    true
+                    false
                 }
             },
-            None => {
-                if self.messages.is_empty() {
-                    false
-                } else {
-                    self.set_active_message(0);
-                    true
-                }
-            }
+            None => self.first_message()
         }
     }
 
-    pub fn first_message(&mut self) {
-        self.set_active_message(0);
+    pub fn first_message(&mut self) -> bool {
+        self.set_active_message(self.display_messages.first().unwrap().0)
     }
 
-    pub fn last_message(&mut self) {
-        self.set_active_message(self.messages.len() - 1);
+    pub fn last_message(&mut self) -> bool {
+        self.set_active_message(self.display_messages.last().unwrap().0)
     }
 }
 
-impl<MessageType: ProtobufMessageEnumTraits + ToString + Clone + 'static> ViewModel for ProtobufMessageListViewModel<MessageType> {
+impl<MessageType: ProtobufMessageEnumTraits + Clone + 'static> ViewModel for ProtobufMessageListViewModel<MessageType> {
     fn draw(&mut self, ui: &mut eframe::egui::Ui, events: &mut Vec<Event>) {
 
         let avail_space = ui.available_size();
@@ -158,15 +238,59 @@ impl<MessageType: ProtobufMessageEnumTraits + ToString + Clone + 'static> ViewMo
                     ui.set_width(message_list_width);
                     ui.set_height(avail_space.y - table_constants::BOTTOM_MARGIN);
 
-                    let mut table_builder = TableBuilder::new(ui);
+                    if self.filterable {
+                        ui.with_layout(
+                            Layout::right_to_left(Align::TOP),
+                            |ui| {
+                                ui.set_width(message_list_width);
+                                ui.add_space(20.0);
 
-                    if self.b_scroll_next {
-                        table_builder = table_builder.scroll_to_row
-                            (self.active_message.unwrap_or(0), None);
-                        self.b_scroll_next = false;
-                    }
+                                if egui::ComboBox::new(
+                                    ui.next_auto_id(),
+                                    "Filter"
+                                ).width(200.0)
+                                .show_index(
+                                    ui,
+                                    &mut self.active_filter_index,
+                                    self.filterable_list.len(),
+                                    |i| self.filterable_list[i].clone()
+                                )
+                                .changed() {
+                                    self.display_messages = {
+                                        if self.active_filter_index != 0 {
+                                            let msg_str = self.filterable_data[self.active_filter_index - 1].0;
+                                            self.messages[msg_str].clone().into_iter().collect()
+                                        } else {
+                                            self.messages["None"].clone().into_iter().collect()
+                                        }
+                                    };
+                                    self.first_message();
+                                }
+                            });
+                        }
 
-                    table_builder = table_builder.striped(true)
+                        ui.end_row();
+                        let mut table_builder = TableBuilder::new(ui);
+
+                        if self.b_scroll_next {
+                            // get real index
+                            let a_index = self.active_message.unwrap_or(0);
+                            let mut real_index = 0;
+                            for i in 0..self.display_messages.len() {
+                                let msg = &self.display_messages[i];
+                                if a_index == msg.0 {
+                                    real_index = i;
+                                    break
+                                }
+                            }
+                            table_builder = table_builder.scroll_to_row(
+                                real_index,
+                                None
+                            );
+                            self.b_scroll_next = false;
+                        }
+
+                        table_builder = table_builder.striped(true)
                     .column(Column::exact(table_constants::COL_INDEX_WIDTH));
 
                     if self.message_ticks.is_some() {
@@ -193,19 +317,24 @@ impl<MessageType: ProtobufMessageEnumTraits + ToString + Clone + 'static> ViewMo
                         });
                     })
                     .body(|body| {
-                        body.rows(table_constants::ROW_HEIGHT, self.messages.len(), |index, mut row| {
-                            let message_return = &self.messages[index];
+                        body.rows(
+                            table_constants::ROW_HEIGHT,
+                            self.display_messages.len(),
+                            |index, mut row| {
+
+                            let message_return = &self.display_messages[index];
                             let name = {
-                                if message_return.message.is_some() {
-                                    message_return.message.as_ref().unwrap().to_string()
+                                if message_return.1.message.is_some() {
+                                    message_return.1.message.as_ref().unwrap().to_str()
                                 } else {
                                     "????".into()
                                 }
                             };
-
+                            
+                            let real_index = self.display_messages[index].0;
                             let is_active = {
                                 if let Some(active_index) = self.active_message {
-                                    if index == active_index {
+                                    if real_index == active_index {
                                         true
                                     } else {
                                         false
@@ -217,7 +346,7 @@ impl<MessageType: ProtobufMessageEnumTraits + ToString + Clone + 'static> ViewMo
 
                             let mut responses = Vec::new();
                             responses.push(row.col(|ui| {
-                                let msg = format!("{}", index + 1);
+                                let msg = format!("{}", real_index + 1);
                                 if is_active {
                                     ui.label(RichText::new(msg).color(table_constants::SELECTED_ITEM_COLOUR));
                                 } else {
@@ -226,7 +355,7 @@ impl<MessageType: ProtobufMessageEnumTraits + ToString + Clone + 'static> ViewMo
                             }).1);
 
                             if let Some(ticks) = &self.message_ticks {
-                                let tick = ticks[index];
+                                let tick = ticks[real_index];
 
                                 responses.push(row.col(|ui| {
                                     let tick = tick.to_string();
@@ -284,7 +413,7 @@ impl<MessageType: ProtobufMessageEnumTraits + ToString + Clone + 'static> ViewMo
                     if let Some(pm_vm) = self.vm_protobuf_message.as_mut() {
                         if let Some(msg_header_cb) = &mut self.message_header_callback {
                             let active_index = self.active_message.unwrap();
-                            let msg_ref = &self.messages[active_index];
+                            let msg_ref = &self.messages["None"][&active_index];
                             msg_header_cb(active_index, ui, events, msg_ref);
                         }